Move sys methods into new class/file
